+++ ChangeLog.md
@@ -1,5 +1,6 @@
 # BountyBot Change Log
 ## BountyBot Version 2.12.0fi:
+- Added command `/bounty ping` to mention bounty hunters who have reacted to the bounty's thread or marked themselves interested in the bounty's event
 - `/create-default bounty-board-forum` now has an option for customizing the new channel's name
 ## BountyBot Version 2.11.1ib:
 - Completing or taking down bounties now cancel the bounty's event if it hasn't been closed already